requirements.txt
accessible-pygments==0.0.5
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.6.12
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
docutils==0.21.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.56.0
green==4.0.2
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
humanize==4.12.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
ipyvue==1.11.2
ipyvuetify==1.11.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-cache==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
loguru==0.7.3
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
Mesa==3.1.4
Mesa-Viz-Tornado==0.1.3
myst-nb==1.2.0
myst-parser==4.0.1
narwhals==1.28.0
nbclient==0.10.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
pydata-sphinx-theme==0.16.1
pydot==3.0.4
Pygments==2.19.1
pymdown-extensions==10.14.3
pyparsing==3.2.1
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-statemachine==2.5.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
reacton==1.9.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rich-click==1.8.6
roman-numerals-py==3.1.0
rpds-py==0.23.1
ruff==0.9.7
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
solara==1.44.1
solara-server==1.44.1
solara-ui==1.44.1
soupsieve==2.6
Sphinx==8.2.1
sphinx-copybutton==0.5.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.0
tabulate==0.9.0
termcolor==2.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0
wheel==0.45.1
widgetsnbextension==4.0.13
zipp==3.21.0
